move setup into utils
